Move dispersion fitter from plugins to tidy3d
