+++ docs/api/plugins/dispersion.rst
@@ -1,16 +1,13 @@
 
 .. currentmodule:: tidy3d
 
-Dispersive Model Fitting
-------------------------
+Remote Dispersive Model Fitting
+-------------------------------
 
 .. autosummary::
    :toctree: ../_autosummary/
    :template: module.rst
 
-   tidy3d.plugins.dispersion.FastDispersionFitter
-   tidy3d.plugins.dispersion.AdvancedFastFitterParam
-   tidy3d.plugins.dispersion.DispersionFitter
    tidy3d.plugins.dispersion.AdvancedFitterParam
    tidy3d.plugins.dispersion.web.run
    tidy3d.plugins.dispersion.StableDispersionFitter
