+++ tidy3d/components/fitter/fit_fast.py
@@ -9,12 +9,12 @@ import scipy
 from pydantic.v1 import Field, NonNegativeFloat, PositiveFloat, PositiveInt, validator
 from rich.progress import Progress
 
-from ...components.base import Tidy3dBaseModel, cached_property, skip_if_fields_missing
-from ...components.medium import LOSS_CHECK_MAX, LOSS_CHECK_MIN, LOSS_CHECK_NUM, PoleResidue
-from ...components.types import ArrayComplex1D, ArrayComplex2D, ArrayFloat1D, ArrayFloat2D
 from ...constants import C_0
 from ...exceptions import ValidationError
 from ...log import get_logging_console, log
+from ..base import Tidy3dBaseModel, cached_property, skip_if_fields_missing
+from ..medium import LOSS_CHECK_MAX, LOSS_CHECK_MIN, LOSS_CHECK_NUM, PoleResidue
+from ..types import ArrayComplex1D, ArrayComplex2D, ArrayFloat1D, ArrayFloat2D
 from .fit import DispersionFitter
 
 # numerical tolerance for pole relocation for fast fitter
